backend/requirements.txt
Flask
flask_sqlalchemy
flask_cors
sentence-transformers
psycopg2-binary
openai
